# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlexScale",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api",
      "variables": {}
    }
  ],
  "paths": {
...
    "/document": {
      "post": {
        "tags": [
          "Document"
        ],
        "summary": "uploadDoc",
        "operationId": "uploadDoc",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/uploadDocRequest"
                  },
                  {
                    "example": {
                      "Location": "https://flexscalebucket.s3-us-west-2.amazonaws.com/uploads/63b6ad27e6aad9c1d9c621b8/Test.png",
                      "User": "63b6ad27e6aad9c1d9c621b8",
                      "prevDocIds": [
                        "63bbc09cf7fa506de853b795"
                      ],
                      "FileName": "Test",
                      "CreatedBy": "6386106b8534e766fcbada14"
                    }
                  }
                ]
              },
              "example": {
                "Location": "https://flexscalebucket.s3-us-west-2.amazonaws.com/uploads/63b6ad27e6aad9c1d9c621b8/Test.png",
                "User": "63b6ad27e6aad9c1d9c621b8",
                "prevDocIds": [
                  "63bbc09cf7fa506de853b795"
                ],
                "FileName": "Test",
                "CreatedBy": "6386106b8534e766fcbada14"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "put": {
        "tags": [
          "Document"
        ],
        "summary": "editDoc",
        "operationId": "editDoc",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63bbcbf1f7fa506de853c357"
            }
          },
          {
            "name": "FileName",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "Test Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "delete": {
        "tags": [
          "Document"
        ],
        "summary": "deleteDoc",
        "operationId": "deleteDoc",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63bbcbf1f7fa506de853c357"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      }
    }
  },
  "components": {
    "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
    },
    "schemas": {
...
      "uploadDocRequest": {
        "title": "uploadDocRequest",
        "required": [
          "Location",
          "User",
          "prevDocIds",
          "FileName",
          "CreatedBy"
        ],
        "type": "object",
        "properties": {
          "Location": {
            "type": "string"
          },
          "User": {
            "type": "string"
          },
          "prevDocIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": ""
          },
          "FileName": {
            "type": "string"
          },
          "CreatedBy": {
            "type": "string"
          }
        },
        "example": {
          "Location": "https://flexscalebucket.s3-us-west-2.amazonaws.com/uploads/63b6ad27e6aad9c1d9c621b8/Test.png",
          "User": "63b6ad27e6aad9c1d9c621b8",
          "prevDocIds": [
            "63bbc09cf7fa506de853b795"
          ],
          "FileName": "Test",
          "CreatedBy": "6386106b8534e766fcbada14"
        }
      }
    }
  },
...
}